chore: Move all CI to Azure Pipelines (#855)
It's easier to just have one CI for all of the CI builds for Appium Desktop. Opted for Azure Pipelines because it supports all 3 platforms.

Have one build that runs the packager on all 3 platforms and one build that does the crowdin sync.

Took out Greenkeeper lockfile stuff because that's not needed anymore. Greenkeeper does it for you now.

Removed references to AppVeyor and Travis.