fix: Stop ignoring result for element location/size (#922)
The 'ignoreResult' flag was causing the result to not be sent back to the renderer process. This was done because there's some cases where we don't want the result to be displayed using "notification" but this is a problem for highlighting an element because it wasn't getting the result.

This changes it so that 'ignoreResult' still sends the result back to the renderer but skips the notification if it's 'ignoreResult=true'.